+++ app/controllers/templates_controller.rb
@@ -5,6 +5,8 @@ class TemplatesController < ApplicationController
   before_action :authenticate_user!
   before_action :set_template, only: [:show, :edit, :update, :destroy]
 
+  layout "layouts/full_page_sidebar"
+
   # # POST /template/add_item.js
   # def add_item
   # end