+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **General Changes**
   - Update interface for setting up accounts, charge types, and templates
   - Improve overview graph year selection
+  - Entry dates now link to the appropriate month on the calendar
 - **Gem Changes**
   - Update to ruby 2.6.1
   - Update to rails 6.0.0.beta1